[all services] performance : Inter service commiunication with openfeign client

+++ upload-service/src/main/java/ai/deepdetect/validators/impl/FileSizeLimitValidator.java
@@ -8,7 +8,7 @@ import jakarta.validation.ConstraintValidatorContext;
 
 public class FileSizeLimitValidator implements ConstraintValidator<FileSizeLimit, MultipartFile>  {
 
-    private long size;
+    private double size;
 
     @Override
     public void initialize(FileSizeLimit constraintAnnotation) {
